refactor - 배틀 진영 선택 제출 로직을 useTeamSelectSubmit 훅으로 분리
- OAuth 사용자와 게스트 로그인 후 배틀 스토어 초기화 및 배틀 페이지 이동 로직 캡슐화
