+++ frontend/src/pages/teamSelectPage/index.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
-import { selectIsLoggingIn, selectIsOAuth, selectUser, useAuthStore } from '@/commons/stores/authStore';
-import { useNavigate, useParams } from 'react-router-dom';
+import { selectIsLoggingIn, useAuthStore } from '@/commons/stores/authStore';
+import { useParams } from 'react-router-dom';
 import { useStepFlow } from './hooks/useStepFlow';
+import { useTeamSelectSubmit } from './hooks/useTeamSelectSubmit';
 import { useGetBattleInfo } from '@/commons/hooks/useGetBattleInfo';
-import { useBattleStore } from '@/pages/battlePage/stores/battleStore';
 import StepIndicator from './components/StepIndicator';
 import StepNavigation from './components/StepNavigation';
 import TeamSelectHeader from './components/TeamSelectHeader';
@@ -11,51 +11,18 @@ import StepContent from './components/StepContent';
 import type { Team } from '@/commons/types/battle';
 
 export default function TeamSelectPage() {
-  const navigate = useNavigate();
   const { id } = useParams<{ id: string }>();
   const { battleInfo } = useGetBattleInfo(id!);
   const hasReferenceData = !!battleInfo?.referenceData;
   const totalSteps = hasReferenceData ? 5 : 4;
   const { currentStep, goToNext, goToPrev, canGoNext } = useStepFlow({ totalSteps });
   const [selectedTeam, setSelectedTeam] = useState<Team | null>(null);
-  const loginGuest = useAuthStore((s) => s.loginGuest);
   const isLoggingIn = useAuthStore(selectIsLoggingIn);
-  const user = useAuthStore(selectUser);
-  const isOAuth = useAuthStore(selectIsOAuth);
+  const { handleSubmit } = useTeamSelectSubmit({ battleId: id!, selectedTeam });
 
   if (!battleInfo) {
     return null;
   }
-
-  const handleSubmit = async () => {
-    if (selectedTeam && id) {
-      // OAuth 사용자는 바로 배틀 페이지로 이동
-      if (user && isOAuth) {
-        useBattleStore.getState().initializeBattle({
-          userId: user.id,
-          battleId: id
-        });
-        useBattleStore.getState().setSelectedTeam(selectedTeam);
-        navigate(`/battle/${id}`, { state: { selectedTeam } });
-        return;
-      }
-
-      // 비회원이거나 로그인 안 된 경우 서버에서 랜덤 닉네임 생성 후 로그인
-      try {
-        const guestUser = await loginGuest(id, selectedTeam !== 'NONE' ? selectedTeam : undefined);
-
-        useBattleStore.getState().initializeBattle({
-          userId: guestUser.id,
-          battleId: id
-        });
-
-        useBattleStore.getState().setSelectedTeam(selectedTeam);
-        navigate(`/battle/${id}`, { state: { selectedTeam } });
-      } catch (e) {
-        alert(e instanceof Error ? e.message : '로그인에 실패했습니다.');
-      }
-    }
-  };
 
   return (
     <main className="min-h-screen bg-[#0a0a1a] py-12 px-4">
